refactor(LottoService): LottoInfo 클래스에서 LottoService 클래스로 수정
LottoService
- 로또 당첨번호, 보너스 번호를 가짐
